moved unit tests and updated to keep up with basic_cli
