+++ Cargo.toml
@@ -14,6 +14,10 @@ keywords = ["Ethereum",
 license = "Unlicense"
 repository = "https://github.com/Drops-of-Diamond/diamond_drops"
 
+[lib]
+name = "diamond_drops"
+path = "src/lib.rs"
+
 [dependencies]
 ethereum-types = "0.3.0"
 ethcore-bytes = "0.1.0"
