--- tests/smc_listener_test.rs
@@ -1,9 +1,0 @@
-extern crate diamond_drops;
-
-use diamond_drops::smc_listener;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-}
